+++ ax3l/constants/DAx3l.py
@@ -1,7 +1,7 @@
 from typing import Final
 
 class DAx3l:
-    VERSION: Final[str] = "0.4.1"
+    VERSION: Final[str] = "0.5.0"
 
     BASE_DIR: Final[str] = "/opt/prod/ax3l"
     BASE_DIR_QA: Final[str] = "/opt/qa/ax3l"